When building input layouts, put VS element hash to the upper 32 bits to reduce possibility of collisions.

+++ Source/Urho3D/Graphics/Direct3D11/D3D11ShaderVariation.h
@@ -150,7 +150,7 @@ private:
     /// Shader type.
     ShaderType type_;
     /// Vertex element hash for vertex shaders. Zero for pixel shaders. Note that hashing is different than vertex buffers.
-    unsigned elementHash_;
+    unsigned long long elementHash_;
     /// Shader parameters.
     HashMap<StringHash, ShaderParameter> parameters_;
     /// Texture unit use flags.
